Build smallc
smallc is an implementation of the C standard library which leverages
musl-libc (https://git.musl-libc.org/git/musl). This patch builds
smallc.
