+++ deps/rust-tpm-20-ref/smallc/include/.gitignore
@@ -1,0 +1,4 @@
+# Ignore everything in this directory
+*
+# Except this file
+!.gitignore
