--- deps/rust-tpm-20-ref/smallc/include/assert.h
@@ -1,8 +1,0 @@
-#ifndef _ASSERT_H_
-#define _ASSERT_H_
-
-#define assert(x) ((void)((x) || (__assert_fail(#x, __FILE__, __LINE__, __func__), 0)))
-
-void __assert_fail(const char *, const char *, int, const char *);
-
-#endif
