--- deps/rust-tpm-20-ref/smallc/include/ctype.h
@@ -1,17 +1,0 @@
-#ifndef _CTYPE_H
-#define _CTYPE_H
-int isalnum(int);
-int isalpha(int);
-int isblank(int);
-int iscntrl(int);
-int isdigit(int);
-int isgraph(int);
-int islower(int);
-int isprint(int);
-int ispunct(int);
-int isspace(int);
-int isupper(int);
-int isxdigit(int);
-int tolower(int);
-int toupper(int);
-#endif
