--- deps/rust-tpm-20-ref/smallc/include/strings.h
@@ -1,9 +1,0 @@
-#ifndef _STRING_H_
-#define _STRING_H_
-
-#include <stddef.h>
-
-int strcasecmp(const char *, const char *);
-int strncasecmp(const char *, const char *, size_t);
-
-#endif
