+++ deps/rust-tpm-20-ref/smallc/src/legacy/err.c
@@ -1,0 +1,90 @@
+/* Copyright (C) 2011 by Valentin Ochs
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
+ */
+
+/*
+ Copyright (c) 2022 - 2023 Intel Corporation
+ SPDX-License-Identifier: Apache-2.0
+*/
+
+#include <err.h>
+#include <stdio.h>
+#include <stdarg.h>
+#include <stdlib.h>
+
+extern char *__progname;
+
+int errno = 0;
+
+void vwarn(const char *fmt, va_list ap)
+{
+  printf ("%s: ", __progname);
+  if (fmt) {
+    printf(fmt, ap);
+  }
+}
+
+void vwarnx(const char *fmt, va_list ap)
+{
+  printf ("%s: ", __progname);
+  if (fmt) printf(fmt, ap);
+}
+
+_Noreturn void verr(int status, const char *fmt, va_list ap)
+{
+  vwarn(fmt, ap);
+}
+
+_Noreturn void verrx(int status, const char *fmt, va_list ap)
+{
+  vwarnx(fmt, ap);
+}
+
+void warn(const char *fmt, ...)
+{
+  va_list ap;
+  va_start(ap, fmt);
+  vwarn(fmt, ap);
+  va_end(ap);
+}
+
+void warnx(const char *fmt, ...)
+{
+  va_list ap;
+  va_start(ap, fmt);
+  vwarnx(fmt, ap);
+  va_end(ap);
+}
+
+_Noreturn void err(int status, const char *fmt, ...)
+{
+  va_list ap;
+  va_start(ap, fmt);
+  verr(status, fmt, ap);
+  va_end(ap);
+}
+
+_Noreturn void errx(int status, const char *fmt, ...)
+{
+  va_list ap;
+  va_start(ap, fmt);
+  verrx(status, fmt, ap);
+  va_end(ap);
+}
