+++ deps/rust-tpm-20-ref/smallc/src/malloc/mallocng/free.c
@@ -1,0 +1,13 @@
+/*
+ Copyright (c) 2022 - 2023 Intel Corporation
+ SPDX-License-Identifier: Apache-2.0
+*/
+
+#include <stdlib.h>
+
+extern __fw_free(void *p);
+
+void free(void *p)
+{
+  __fw_free(p);
+}
