--- deps/rust-tpm-20-ref/smallc/src/stdlib/__need_impl_abort.c
@@ -1,8 +1,0 @@
-#include <stdlib.h>
-
-// TBD:
-extern void __fw_abort();
-
-void abort(void) {
-    __fw_abort();
-}
