--- deps/rust-tpm-20-ref/smallc/src/string/strings.c
@@ -1,17 +1,0 @@
-#include <strings.h>
-#include <ctype.h>
-
-int strcasecmp(const char *_l, const char *_r)
-{
-	const unsigned char *l=(void *)_l, *r=(void *)_r;
-	for (; *l && *r && (*l == *r || tolower(*l) == tolower(*r)); l++, r++);
-	return tolower(*l) - tolower(*r);
-}
-
-int strncasecmp(const char *_l, const char *_r, size_t n)
-{
-	const unsigned char *l=(void *)_l, *r=(void *)_r;
-	if (!n--) return 0;
-	for (; *l && *r && n && (*l == *r || tolower(*l) == tolower(*r)); l++, r++, n--);
-	return tolower(*l) - tolower(*r);
-}
